Перестроены модели. Контроллеры переписаны под Entity Framework Core. Убраны ненужные мапперы. Добавлены данные в dbContext. Переделаны Dto классы.

+++ EnterpriseWarehouse/EnterpriseWarehouseServer/Controllers/InvoiceController.cs
@@ -1,131 +1,170 @@
-﻿using Enterprise.Data;
+﻿using AutoMapper;
+using Enterprise.Data;
 using EnterpriseWarehouseServer.Dto;
-using EnterpriseWarehouseServer.Repositories;
 using Microsoft.AspNetCore.Mvc;
-
+using Microsoft.EntityFrameworkCore;
 
 namespace EnterpriseWarehouseServer.Controllers;
 [Route("api/[controller]")]
 [ApiController]
 public class InvoiceController : ControllerBase
 {
-    private readonly ILogger<InvoiceController> _logger;
+    private readonly ILogger<ProductController> _logger;
 
-    private readonly IMainRepository _context;
-    public InvoiceController(ILogger<InvoiceController> logger, IMainRepository mainRepository)
+    private readonly EnterpriseWarehouseDbContext _context;
+
+    private readonly IMapper _mapper;
+    public InvoiceController(ILogger<ProductController> logger, EnterpriseWarehouseDbContext context, IMapper mapper)
     {
         _logger = logger;
-        _context = mainRepository;
+        _context = context;
+        _mapper = mapper;
     }
 
     /// <summary>
     ///     [HttpGet] - return all invoice
     /// </summary>
-    /// <returns>List of Invoice</returns>
+    /// <returns>List of Invoices</returns>
     [HttpGet]
-    public IEnumerable<InvoiceGetDto> Get()
+    public async Task<ActionResult<IEnumerable<InvoiceGetDto>>> Get()
     {
         _logger.LogInformation("Get invoice.");
-        var result = (from invoice in _context.Invoices
-                      group invoice by new
-                      {
-                          invoiceId = invoice.Id,
-                          nameOrg = invoice.NameOrganization,
-                          addressOrg = invoice.AdressOrganization,
-                          shipmentDate = invoice.ShipmentDate
-                      } into grp
-                      select new InvoiceGetDto
-                      {
-                          Id = grp.Key.invoiceId,
-                          NameOrganization = grp.Key.nameOrg,
-                          AdressOrganization = grp.Key.addressOrg,
-                          ShipmentDate = grp.Key.shipmentDate.ToShortDateString(),
-                          Products = (from invoiceContent in _context.InvoicesContent
-                                      where invoiceContent.InvoiceId == grp.Key.invoiceId
-                                      select (invoiceContent.ProductItemNumber, invoiceContent.Quantity)).ToDictionary(x => x.ProductItemNumber, x => x.Quantity)
-                      });
-        return result;
-        /* return _context.Invoices.Select(invoice =>
-             new InvoiceGetDto
-             {
-                 Id = invoice.Id,
-                 NameOrganization = invoice.NameOrganization,
-                 AdressOrganization = invoice.AdressOrganization,
-                 ShipmentDate = invoice.ShipmentDate.ToShortDateString(),
-                 Products = invoice.InvoiceContentId,
-             }
-         );*/
+        if (_context.Invoices != null && _context.Products != null)
+        {
+            var result_temp = await (from invoice in _context.Invoices
+                                     group invoice by new
+                                     {
+                                         invoiceId = invoice.Id,
+                                         nameOrg = invoice.NameOrganization,
+                                         addressOrg = invoice.AddressOrganization,
+                                         shipmentDate = invoice.ShipmentDate
+                                     } into grp
+                                     select new
+                                     {
+                                         Id = grp.Key.invoiceId,
+                                         NameOrganization = grp.Key.nameOrg,
+                                         AdressOrganization = grp.Key.addressOrg,
+                                         ShipmentDate = grp.Key.shipmentDate.ToShortDateString()
+                                     }).ToListAsync();
+            var result = new List<InvoiceGetDto>();
+            foreach (var item in result_temp)
+            {
+                var products = await (from invoice in _context.Invoices
+                                      from invoiceContent in invoice.InvoicesContent
+                                      join product in _context.Products on invoiceContent.ProductItemNumber equals product.ItemNumber
+                                      where invoice.Id == item.Id
+                                      select new { invoiceContent.ProductItemNumber, invoiceContent.Quantity }).ToDictionaryAsync(x => x.ProductItemNumber, x => x.Quantity);
+                result.Add(new InvoiceGetDto
+                {
+                    Id = item.Id,
+                    NameOrganization = item.NameOrganization,
+                    AdressOrganization = item.AdressOrganization,
+                    ShipmentDate = item.ShipmentDate,
+                    Products = products
+                });
+            }
+            return result;
+        }
+        else
+            return NotFound();
     }
 
     /// <summary>
     ///     [HttpGet("{id}")] - return invoice with id
     /// </summary>
-    /// <param Id = "id" >Id of the Invoice to be view</param>
+    /// <param Id = "id" >Id of the Invoices to be view</param>
     /// <returns>Info of product</returns>
     [HttpGet("{id}")]
-    public ActionResult<InvoiceGetDto?> Get(int id)
+    public async Task<ActionResult<InvoiceGetDto?>> Get(int id)
     {
-        var invoice = _context.Invoices.FirstOrDefault(invoice => invoice.Id == id);
-        if (invoice == null)
+        if (_context.Invoices != null && _context.InvoicesContent != null)
         {
-            _logger.LogInformation("Not found ivoice with {id}.", id);
-            return NotFound();
+            var invoice = await _context.Invoices.FirstOrDefaultAsync(invoice => invoice.Id == id);
+            if (invoice == null)
+            {
+                _logger.LogInformation("Not found ivoice with {id}.", id);
+                return NotFound();
+            }
+            else
+            {
+                _logger.LogInformation("Get invoice with {id}.", id);
+                return Ok(new InvoiceGetDto
+                {
+                    Id = invoice.Id,
+                    NameOrganization = invoice.NameOrganization,
+                    AdressOrganization = invoice.AddressOrganization,
+                    ShipmentDate = invoice.ShipmentDate.ToShortDateString(),
+                    Products = _context.InvoicesContent.Where(res => res.InvoiceId == id).Select(result =>
+                    new { result.ProductItemNumber, result.Quantity }).ToDictionary(x => x.ProductItemNumber, x => x.Quantity)
+                });
+            }
         }
         else
         {
-            _logger.LogInformation("Get invoice with {id}.", id);
-            return Ok(new InvoiceGetDto
-            {
-                Id = invoice.Id,
-                NameOrganization = invoice.NameOrganization,
-                AdressOrganization = invoice.AdressOrganization,
-                ShipmentDate = invoice.ShipmentDate.ToShortDateString(),
-                Products = _context.InvoicesContent.Where(res => res.InvoiceId == id).Select(result =>
-                (result.ProductItemNumber, result.Quantity)).ToDictionary(x => x.ProductItemNumber, x => x.Quantity)
-            });
+            _logger.LogInformation("Not found invoice with {id}.", id);
+            return NotFound();
         }
     }
 
     /// <summary>
     ///     [HttpPost] - add new invoice
     /// </summary>
-    /// <param Invoice>Add new Invoice</param>
+    /// <param Invoices>Add new Invoices</param>
+    /// <returns>Result of operation</returns>
     [HttpPost]
-    public void Post([FromBody] InvoicePostDto invoice)
+    [ProducesResponseType(201)]
+    public async Task<ActionResult<Invoice>> Post([FromBody] InvoicePostDto invoice)
     {
-        var new_invoice = new Invoice(
-            invoice.Id,
-            invoice.NameOrganization,
-            invoice.AdressOrganization,
-            DateTime.ParseExact(invoice.ShipmentDate, "dd.mm.yyyy", null).Date
-            );
-        foreach (var elem in invoice.Products)
+        if (_context.Invoices != null)
         {
-            var product = _context.Products.FirstOrDefault(x => x.ItemNumber == elem.Key);
-            var new_invoiceContent = new InvoiceContent(
-                 (uint)new Random().Next(4, 10000),
-                 new_invoice,
-                 product,
-                 elem.Value
+            var new_invoice = new Invoice(
+                invoice.Id,
+                invoice.NameOrganization,
+                invoice.AdressOrganization,
+                DateTime.Parse(invoice.ShipmentDate).Date
                 );
-            new_invoiceContent.ProductItemNumber = elem.Key;
-            new_invoiceContent.InvoiceId = invoice.Id;
-            new_invoice.InvoiceContent.Add(new_invoiceContent);
-            product.InvoiceContent.Add(new_invoiceContent);
-            _context.InvoicesContent.Add(new_invoiceContent);
+            var mappedInvoice = _mapper.Map<Invoice>(new_invoice);
+            _context.Invoices.Add(mappedInvoice);
+            foreach (var elem in invoice.Products)
+            {
+                if (_context.Products != null && _context.InvoicesContent != null)
+                {
+                    var product = await _context.Products.FirstOrDefaultAsync(x => x.ItemNumber == elem.Key);
+                    if (product != null)
+                    {
+                        var new_invoiceContent = new InvoiceContent
+                        {
+                            InvoiceId = new_invoice.Id,
+                            ProductItemNumber = product.ItemNumber,
+                            Quantity = elem.Value
+                        };
+                        var mappedInvoiceContent = _mapper.Map<InvoiceContent>(new_invoiceContent);
+                        _context.InvoicesContent.Add(mappedInvoiceContent);
+                    }
+                    else
+                        return Problem("Not found product with item number.");
+                }
+                else
+                    return Problem("Entity set 'EnterpriseWarehouseDbContext.Products is null.");
+            }
+            _context.SaveChanges();
+            return CreatedAtAction("Post", new { id = invoice.Id }, _mapper.Map<InvoiceGetDto>(invoice));
         }
-        _context.Invoices.Add(new_invoice);
+        else
+            return Problem("Entity set 'EnterpriseWarehouseDbContext.Invoices is null.");
     }
 
     /// <summary>
     ///     [HttpPut("{id}")] - update info of invoice with id
     /// </summary>
-    /// <param Id="id">Id of the Invoice to be update</param>
+    /// <param Id="id">Id of the Invoices to be update</param>
     /// <returns>Result of operation</returns>
     [HttpPut("{id}")]
-    public ActionResult<InvoicePostDto?> Put(int id, [FromBody] InvoicePostDto invoice_)
+    public async Task<IActionResult> Put(int id, [FromBody] InvoicePostDto invoice_)
     {
-        var invoice = _context.Invoices.FirstOrDefault(invoice => invoice.Id == id);
+        if (_context.Invoices == null)
+            return NotFound();
+        var invoice = await _context.Invoices.FirstOrDefaultAsync(invoice => invoice.Id == id);
         if (invoice == null)
         {
             _logger.LogInformation("Not found ivoice with {id}.", id);
@@ -136,60 +175,69 @@ public class InvoiceController : ControllerBase
             _logger.LogInformation("Put invoice with {id}.", id);
             invoice.Id = invoice_.Id;
             invoice.NameOrganization = invoice_.NameOrganization;
-            invoice.AdressOrganization = invoice_.AdressOrganization;
+            invoice.AddressOrganization = invoice_.AdressOrganization;
             invoice.ShipmentDate = DateTime.ParseExact(invoice_.ShipmentDate, "dd.mm.yyyy", null).Date;
             foreach (var elem in invoice_.Products)
             {
-                var product = _context.InvoicesContent.FirstOrDefault(product => (product.InvoiceId == invoice.Id && product.ProductItemNumber == elem.Key));
-                if (product != null)
+                if (_context.InvoicesContent != null && _context.Products != null)
                 {
-                    product.Product = _context.Products.FirstOrDefault(_product => _product.ItemNumber == product.ProductItemNumber);
-                    product.InvoiceId = invoice_.Id;
-                    product.ProductItemNumber = elem.Key;
-                    product.Quantity = elem.Value;
+                    var product = await _context.InvoicesContent.FirstOrDefaultAsync(product => (product.InvoiceId == invoice.Id && product.ProductItemNumber == elem.Key));
+                    if (product != null)
+                    {
+                        product.Product = await _context.Products.FirstOrDefaultAsync(_product => _product.ItemNumber == product.ProductItemNumber);
+                        product.InvoiceId = invoice_.Id;
+                        product.ProductItemNumber = elem.Key;
+                        product.Quantity = elem.Value;
+                    }
+                    else
+                        return NotFound();
                 }
                 else
-                {
-                    _context.InvoicesContent.Add(new InvoiceContent(
-                        (uint)new Random().Next(4, 10000),
-                        invoice,
-                        _context.Products.FirstOrDefault(product_ => product_.ItemNumber == elem.Key),
-                        elem.Value
-                    ));
-                }
+                    return Problem("Entity set 'EnterpriseWarehouseDbContext.Products or 'EnterpriseWarehouseDbContext.InvoicesContent is null.");
             }
-
-            return Ok();
+            await _context.SaveChangesAsync();
+            return NoContent();
         }
     }
 
     /// <summary>
     ///     [HttpDelete("{id}")] - delete invoice with id
     /// </summary>
-    /// <param Id="id">Id of the Invoice to be removed</param>
+    /// <param Id="id">Id of the Invoices to be removed</param>
     /// <returns>Result of operation</returns>
     [HttpDelete("{id}")]
-    public ActionResult Delete(int id)
+    public async Task<ActionResult> Delete(int id)
     {
-        var invoice = _context.Invoices.FirstOrDefault(invoice => invoice.Id == id);
-        if (invoice == null)
+        if (_context.Invoices != null)
         {
-            _logger.LogInformation("Not found ivoice with {id}.", id);
-            return NotFound();
+            var invoice = await _context.Invoices.FirstOrDefaultAsync(invoice => invoice.Id == id);
+            if (invoice == null)
+            {
+                _logger.LogInformation("Not found ivoice with {id}.", id);
+                return NotFound();
+            }
+            else
+            {
+                if (_context.InvoicesContent != null)
+                {
+                    _logger.LogInformation("Delete invoice with {id}.", id);
+                    var products = await _context.InvoicesContent.Where(res => res.InvoiceId == id).ToListAsync();
+                    if (products != null)
+                    {
+                        for (var i = 0; i < products.Count(); i++)
+                        {
+                            _context.InvoicesContent.Remove(products[i]);
+                        }
+                    }
+                    _context.Invoices.Remove(invoice);
+                    await _context.SaveChangesAsync();
+                    return NoContent();
+                }
+                else
+                    return Problem("Entity set 'EnterpriseWarehouseDbContext.InvoicesContent is null.");
+            }
         }
         else
-        {
-            _logger.LogInformation("Delete invoice with {id}.", id);
-            var products = _context.InvoicesContent.Where(res => res.InvoiceId == id).ToList();
-            if (products != null)
-            {
-                for (var i = 0; i < products.Count(); i++)
-                {
-                    _context.InvoicesContent.Remove(products[i]);
-                }
-            }
-            _context.Invoices.Remove(invoice);
-            return Ok();
-        }
+            return NotFound();
     }
 }
